Split main file into smaller files: models creation and straning, ensemble, optimization, etc. Added optimization- feature selection.
