+++ tests_on_one_ticker.py
@@ -1,13 +1,6 @@
-"""
-Single stock with pipes 
-change from collab trying to run for one stock
-
-Run after setting up an environment - setup.py
-"""
 
 import pandas as pd
 import numpy as np
-#import matplotlib.pyplot as plt
 import seaborn as sns
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import StandardScaler
@@ -21,7 +14,6 @@ import yfinance as yf
 import pandas_ta as ta
 import os
 from sklearn.base import BaseEstimator, TransformerMixin
-# from datetime import datetime
 from pandas_datareader import data as pdr
 import datetime
 
@@ -34,309 +26,6 @@ current_date = datetime.datetime.now().strftime("%Y%m%d")
 # Create date folder inside results
 date_folder = f'results/{current_date}'
 os.makedirs(date_folder, exist_ok=True)
-
-# ===============================================================================
-# Data Pipeline Classes
-## Pipeline to fetch stock data and feature engineering
-# ===============================================================================
-# fetch historical stock data
-class DataFetcher(BaseEstimator, TransformerMixin):
-    def __init__(self, ticker_symbol, start_date, end_date):
-        self.ticker_symbol = ticker_symbol
-        self.start_date = start_date
-        self.end_date = end_date
-
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-      try:
-        ticker = yf.Ticker(self.ticker_symbol)
-        history = ticker.history(start=self.start_date, end=self.end_date)
-
-        if history.empty:
-          raise ValueError(f"No data returned for ticker: {self.ticker_symbol}")
-
-        return history
-      #    return ticker.history(start=self.start_date, end=self.end_date)
-      except Exception as e:
-        print(f"Error fetching data for ticker {self.ticker_symbol}: {e}")
-
-        return pd.DataFrame() # return an empty df if an error occurs
-
-
-# add market indicators
-class IndicatorCalculator(BaseEstimator, TransformerMixin):
-    def __init__(self, include_prime_rate=True, prime_start="2019-01-01", prime_end="2024-01-01"):
-        self.indicators = [
-            'psar', 'mfi', 'mvp'
-        ]
-        self.include_prime_rate = include_prime_rate
-        self.prime_start = prime_start
-        self.prime_end = prime_end
-        self.prime_data = None
-
-        # Fetch prime rate during initialization if needed
-        if self.include_prime_rate:
-           self._fetch_prime_rate()
-    
-    def _fetch_prime_rate(self):
-       "Fetch prime rate data from FRED"
-       try:
-          start = datetime.datetime.strptime(self.prime_start, "%Y-%m-%d") 
-          end = datetime.datetime.strptime(self.prime_end, "%Y-%m-%d")
-
-          self.prime_data = pdr.get_data_fred('PRIME', start, end)
-          #####MAYBE SHOULD DO FFILL HERE BUT I WANT TO DO IT IN THE SECOND PIPELINE
-          # Forward fill to handle missing dates
-        #   self.prime_data = self.prime_data.fillna(method='ffill').fillna(method='bfill')
-
-          # Step 1: First apply forward fill
-          self.prime_data = self.prime_data.fillna(method='ffill')
-        
-          # Step 2: Then apply backward fill
-          self.prime_data = self.prime_data.fillna(method='bfill')
-          print(f"Prime rate data loaded with {len(self.prime_data)} records")
-       except Exception as e:
-          print(f"Error fetching prime rate data in fetch: {e}")
-          self.prime_data = None
-          
-
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-        data = X.copy()
-        # Calculate technical indicators
-        for indicator in self.indicators:
-            try:
-                if indicator == 'psar':
-                    psar = ta.psar(data['High'], data['Low'], data['Close'])
-                    data['PSAR'] = psar['PSARl_0.02_0.2']
-                elif indicator == 'mfi':
-                    data['MFI'] = ta.mfi(data['High'], data['Low'], data['Close'], data['Volume']).astype(float)
-                # Calculate Moving Volatility Pattern (MVP)
-                # Here we're assuming MVP as a simple moving average of squared returns
-                elif indicator == 'mvp':
-                  data['Returns'] = data['Close'].pct_change()                                                          # calculate the daily return
-                  data['Volatility'] = data['Returns'].rolling(window=28).std() * np.sqrt(252)                          # calculate 28-day rolling volatility (daily annual volatility)
-                  data['MVP'] = data['Volatility'].pct_change()                                                         # the pct change of the volatility, Rate of change of volatility
-                  threshold = 0.01
-                  data['MVP_Signal'] = np.where(abs(data['MVP']) < threshold, 0)                                        # buy when the abs mvp is almost 0
-            except Exception as e:
-                print(f"Error adding indicator {indicator}: {e}")
-            
-            # Add prime rate data
-            if self.include_prime_rate and self.prime_data is not None:
-               try:
-                  # Merge prime rate data with ticker data based on date index
-                  # Make sure the index is datetime for both df
-                  if not isinstance(data.index, pd.DatetimeIndex):
-                     print(f"Warning: Stock data index is not DatetimeIndex. Converting...")
-                     data.index = pd.to_datetime(data.index)
-                  # Add prime rate
-                  data['Prime_Rate'] = None
-                  for date in data.index:
-                     date_str = pd.Timestamp(date)
-                     # Find all prime rated before the current date
-                     applicable_dates = self.prime_data.index[self.prime_data.index <= date_str.to_datetime64()]
-                     # If prime rate was found
-                     if len(applicable_dates) > 0:
-                        closest_date = applicable_dates[-1] # Get the last (most recent) date
-                        data.at[date, 'Prime_Rate'] = self.prime_data.loc[closest_date, 'PRIME']
-
-               except Exception as e:
-                  print(f"Error adding prime rate data: {e}")
-                  
-        return data
-
-# calculate buy and sell signals based on psar indicator
-class SignalCalculator(BaseEstimator, TransformerMixin):
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-        data = X.copy()
-        data['Signal'] = np.where(data['Close'] > data['PSAR'], 1, 0)
-        data['Signal_Shift'] = data['Signal'].shift(1)
-        data['Buy'] = np.where((data['Signal'] == 1) & (data['Signal_Shift'] == 0), data['Close'], np.nan)
-        data['Sell'] = np.where((data['Signal'] == 0) & (data['Signal_Shift'] == 1), data['Close'], np.nan)
-        return data
-
-# for each transaction calculate: volatility, return and sharpe ratio
-class TransactionMetricsCalculator(BaseEstimator, TransformerMixin):
-    def __init__(self, risk_free_rate=0.02):
-        self.risk_free_rate = risk_free_rate
-
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-        data = X.copy()
-
-        # Initialize the transaction metrics columns
-        data['Transaction_Volatility'] = np.nan
-        data['Transaction_Returns'] = np.nan
-        data['Transaction_Sharpe'] = np.nan
-        data['Transaction_Duration'] = np.nan
-        
-        last_buy_index = None
-        last_buy_price = None
-        in_active_position = False
-
-        # Process the data in a single pass
-        for i in range(len(data)):
-            # Record buy signal
-            if not np.isnan(data['Buy'].iloc[i]):
-                last_buy_index = i
-                last_buy_price = data['Close'].iloc[i]
-                in_active_position = True
-
-            # Calculate metrics based on the last known buy signal
-            # (even after a sell signal, until the next buy)
-            if last_buy_index is not None:
-                buy_date = data.index[last_buy_index]
-                current_date = data.index[i]
-                duration = (current_date - buy_date).days
-
-                # Calculate raw return for this transaction
-                returns = (data['Close'].iloc[i] - last_buy_price) / last_buy_price
-
-                # Calculate volatility (standard deviation of daily returns) for this transaction
-                daily_returns = data['Close'].iloc[last_buy_index:i+1].pct_change().dropna()
-                volatility = daily_returns.std() if len(daily_returns) > 1 else 0
-
-                # Calculate Sharpe ratio for this transaction
-                transaction_risk_free_rate = self.risk_free_rate * (duration / 365)
-                sharpe_ratio = (returns - transaction_risk_free_rate) / volatility if volatility != 0 else 0                        # all in the same scale: returns is the return ratio, the vplatility is pct
-
-                # Store the metrics for every day while in a position
-                data.loc[data.index[i], 'Transaction_Volatility'] = volatility
-                data.loc[data.index[i], 'Transaction_Returns'] = returns
-                data.loc[data.index[i], 'Transaction_Sharpe'] = sharpe_ratio
-                data.loc[data.index[i], 'Transaction_Duration'] = duration
-
-                # If this is a sell day, mark that we're no longer in an active position
-                # but continue calculating using the same buy reference point
-                if not np.isnan(data['Sell'].iloc[i]):
-                    in_active_position = False
-
-        return data
-    
-
-# ===============================================================================
-# Data Cleaning Classes
-## Pineline to clean the data, remove outliers, and handle correlation
-# ===============================================================================
-# use forward fill then backward filll (normal in financial data) on all continues columns
-class MissingValueHandler(BaseEstimator, TransformerMixin):
-  def __init__(self, fill_columns = ['Open', 'High', 'Low', 'Close', 'Volume', 'PSAR', 'MFI', 'MVP']):
-    self.fill_columns = fill_columns
-
-  def fit(self, X, y=None):
-    return self
-
-  def transform(self, X):
-    X_ = X.copy()
-    for col in self.fill_columns:
-      if col in X_.columns:
-        X_[col] = X_[col].fillna(method='ffill').fillna(method='bfill')
-
-    # if 'Signal_Shift' in X_.columns:
-    #   X_['Signal_Shift']= X_['Signal_Shift'].fillna(-1)
-    for col in X_.columns:
-      X_[col] = X_[col].fillna(-1)
-
-    return X_
-  
-# handle outliers: Use Z-score to identify data points that deviate significantly from the mean - rolling eondow of 28 days
-class OutlierHandler(BaseEstimator, TransformerMixin):
-  def __init__(self,threshold=3, window=28, exclude_columns=[
-    # Price data
-    'Open', 'High', 'Low', 'Close',
-    # Signal columns
-    'Buy', 'Sell', 'Signal', 'Signal_Shift',
-    # Transaction metrics
-    'Transaction_Sharpe', 'Transaction_Returns', 'Transaction_Volatility', 'Transaction_Duration',
-    # External data
-    'Prime_Rate'
-    ]):
-     
-    self.threshold = threshold
-    self.window = window
-    self.exclude_columns = exclude_columns
-
-  def fit(self, X, y=None):
-    return self
-
-  def transform(self, X):
-    X_ = X.copy()
-    X_ = X_.dropna(axis =1, how = 'all')
-    for column in X_.columns:
-      if column in self.exclude_columns:
-        continue
-      
-      if X_[column].dtype in ['float64', 'int64']:
-        rolling_mean = X_[column].rolling(window=self.window).mean()
-        rolling_std = X_[column].rolling(window=self.window).std()
-        z_scores = np.abs((X_[column] - rolling_mean) / rolling_std)
-        X_[column] = X_[column].mask(z_scores > self.threshold, rolling_mean)
-
-    return X_
-  
-# Handle highly correlated features
-class CorrelationHandler(BaseEstimator, TransformerMixin):
-  def __init__(self, threshold=0.95, keep_columns=['Close']):
-    self.threshold = threshold
-    self.columns_to_drop = None
-    self.keep_columns = keep_columns
-
-  def fit(self, X, y=None):
-    # handle high correlation of continues columns
-    corr_matrix = X.corr().abs()
-    upper = corr_matrix.where(np.triu(np.ones(corr_matrix.shape), k=1).astype(bool))
-    self.columns_to_drop = [col for col in upper.columns if any(upper[col] > self.threshold)]
-
-    # check all columns correlation with the target variable, if nan - remove (doesnt help for prediction later)
-    if 'Transaction_Sharpe' in X.columns:
-      target_corr = X.corr()['Transaction_Sharpe']
-      nan_corr_columns = target_corr[target_corr.isna()].index.tolist()
-      self.columns_to_drop.extend(nan_corr_columns)
-    
-    for col in self.keep_columns:
-       if col in self.columns_to_drop:
-          self.columns_to_drop.remove(col)
-
-    if 'Transaction_Sharpe' in self.columns_to_drop:
-      self.columns_to_drop.remove('Transaction_Sharpe')
-
-    return self
-  
-  def transform(self, X):
-    X_ = X.copy()
-    for col in self.columns_to_drop:
-      if col in X_.columns:
-        print(f"Dropping column {col} due to NaN correlation with target")
-    return X_.drop(self.columns_to_drop, axis=1, errors='ignore')
-
-
-# ===============================================================================
-# Create The Pipelines
-# ===============================================================================
-def create_stock_data_pipeline(ticker_symbol, start_date, end_date, risk_free_rate):
-    return Pipeline([
-        ('data_fetcher', DataFetcher(ticker_symbol, start_date, end_date)),
-        ('indicator_calculator', IndicatorCalculator()),
-        ('signal_calculator', SignalCalculator()),
-        ('transaction_metrics_calculator', TransactionMetricsCalculator(risk_free_rate)),
-    ])
-
-def create_data_cleaning_pipeline(correlation_threshold = 0.95):
-  return Pipeline([
-      ('missing_value_handler', MissingValueHandler()),
-      ('outlier_handler', OutlierHandler()),
-      ('correlation_handler', CorrelationHandler(correlation_threshold))
-  ])
 
 
 # ===============================================================================
@@ -439,7 +128,8 @@ def train_lstm_model_with_cv(X_train_val_scaled, Y_train_val, tscv):
 
             # Split the data
             X_train, X_val = X_train_val_scaled[train_idx], X_train_val_scaled[val_idx]
-            Y_train, Y_val = Y_train_val.iloc[train_idx], Y_train_val.iloc[val_idx]
+            #Y_train, Y_val = Y_train_val.iloc[train_idx], Y_train_val.iloc[val_idx]
+            Y_train, Y_val = Y_train_val[train_idx], Y_train_val[val_idx]
 
             # Perform a manual grid search across LSTM parameters
             fold_best_mse = float('inf')
@@ -508,18 +198,6 @@ def train_lstm_model_with_cv(X_train_val_scaled, Y_train_val, tscv):
         }
     
 
-# ===============================================================================
-# Model Creation and Training
-# ===============================================================================
-# Six models we took from stage 1 of the article, each model with list af params to do grid search on (optimize)
-"""
-Support Vector Regression (SVR)
-eXtreme Gradient Boosting (XGBoost)
-Light Gradient Boosting Machine (LightGBM)
-Random Forest regression (RF)
-Gradient Boosting Regression (GBR)
-Long Short Term Memory model (LSTM)
-"""
 
 def create_models():
     try:
@@ -559,365 +237,388 @@ def create_models():
 
 
  # Create all models, split the data - 5 folds, choose best params then train and vlidate each model, save performance results and params
-def train_and_validate_models(X_train_val, Y_train_val):
-  scaler = StandardScaler()                                                      # Add scaling, the features are scaled to have zero mean and unit variance
-  X_train_val_scaled = scaler.fit_transform(X_train_val)
+# def train_and_validate_models(X_train_val, Y_train_val):
+#   # Scale the features
+#   scaler = StandardScaler()                                                      # Add scaling, the features are scaled to have zero mean and unit variance
+#   X_train_val_scaled = scaler.fit_transform(X_train_val)
 
-  tscv = TimeSeriesSplit(n_splits=5)                                              # time series cross-validation with 5 folds, to ensure temporal order (sequence of events in time)
-  models = create_models()
-  results = {}
-
-  for model_name, (model, params_grid) in models.items():                         # iterate each model
-    print(f"Training {model_name}")
-
-    # reshape the data for LSTM model
-    if model_name == 'LSTM':
-      results[model_name] = train_lstm_model_with_cv(X_train_val_scaled, Y_train_val, tscv)
-      continue
-
-    # Add timeout or max iterations for SVR
-    if model_name == 'SVR':
-      model.set_params(max_iter=1000)  # Limit iterations
-
-    best_mse_scores = []                                                               # lists to store scores and parameters for the models
-    best_rmse_scores = []
-    best_params = []
-    best_score = float('inf')
-    best_model = None
-    best_model_prediction = None
-    Y_val_best = None
-
-    for fold, (train_idx, val_idx) in enumerate(tscv.split(X_train_val_scaled)):  # each fold: split the scaled data to train and val (using the tcsv indices)
-      print(f"\nTraining on fold {fold + 1} for {model_name}...")
-      # Split the data
-      X_train, X_val = X_train_val_scaled[train_idx], X_train_val_scaled[val_idx]
-      Y_train, Y_val = Y_train_val.iloc[train_idx], Y_train_val.iloc[val_idx]
+#   # Add target scaling
+#   target_scaler = StandardScaler()
+#   Y_train_val_scaled = target_scaler.fit_transform(Y_train_val.values.reshape(-1, 1)).flatten()
+  
+#   # Save the scaler for later use
+#   # This is important for inverse transformation
+#   global shared_target_scaler
+#   shared_target_scaler = target_scaler
 
 
-      # Find the best params for the current model
-      print(f"Performing GridSearchCV for {model_name} on fold {fold + 1}...")
+#   tscv = TimeSeriesSplit(n_splits=5)                                              # time series cross-validation with 5 folds, to ensure temporal order (sequence of events in time)
+#   models = create_models()
+#   results = {}
 
-      # Reduce the number of CV splits for SVR to speed up
-      cv_splits = 2 if model_name == 'SVR' else 3
+#   for model_name, (model, params_grid) in models.items():                         # iterate each model
+#     print(f"Training {model_name}")
 
-    #   grid_search = GridSearchCV(estimator=model, param_grid=params_grid, scoring='neg_mean_squared_error', cv=cv_splits, n_jobs=-1, verbose=1) # performs search over set of parameters
-      from sklearn.model_selection import GridSearchCV
-      grid_search = GridSearchCV(
-         estimator=model,
-         param_grid=params_grid,
-         scoring='neg_mean_squared_error',
-         cv=cv_splits,
-         n_jobs=-1,
-         verbose=1
-        )
+#     # reshape the data for LSTM model
+#     if model_name == 'LSTM':
+#       results[model_name] = train_lstm_model_with_cv(X_train_val_scaled, Y_train_val_scaled, tscv)
+#       continue
 
-      try:
-        grid_search.fit(X_train, Y_train)                                         # train the model for every combination of parameters on each training set of the fold
-        print(f"Best parameters found for fold {fold + 1}: {grid_search.best_params_}")
+#     # Add timeout or max iterations for SVR
+#     if model_name == 'SVR':
+#       model.set_params(max_iter=1000)  # Limit iterations
 
-        # Save the best params
-        best_params.append(grid_search.best_params_)
+#     best_mse_scores = []                                                               # lists to store scores and parameters for the models
+#     best_rmse_scores = []
+#     best_params = []
+#     best_score = float('inf')
+#     best_model = None
+#     best_model_prediction = None
+#     Y_val_best = None
 
-        # Train the best model
-        best_model_fold = grid_search.best_estimator_                             # the model that performs the best on the last search
-
-        # Validate using the model
-        print("Starting validation: ")
-        Y_pred = best_model_fold.predict(X_val)                                   # use the best model after traind and validate
-
-        # Calculate performance for this fold
-        mse = mean_squared_error(Y_val, Y_pred)
-        rmse = np.sqrt(mse)
-
-        # Save
-        best_mse_scores.append(mse)
-        best_rmse_scores.append(rmse)
-
-        print(f"End of Fold {fold + 1} - MSE: {mse:.4f}, RMSE: {rmse:.4f}")
-
-        # Update best model if this fold's model is better
-        if mse < best_score:
-          best_score = mse
-          best_model = best_model_fold
-          best_model_prediction = Y_pred
-          Y_val_best= Y_val
+#     for fold, (train_idx, val_idx) in enumerate(tscv.split(X_train_val_scaled)):  # each fold: split the scaled data to train and val (using the tcsv indices)
+#       print(f"\nTraining on fold {fold + 1} for {model_name}...")
+#       # Split the data
+#       X_train, X_val = X_train_val_scaled[train_idx], X_train_val_scaled[val_idx]
+#       #Y_train, Y_val = Y_train_val_scaled.iloc[train_idx], Y_train_val_scaled.iloc[val_idx]
+#       Y_train, Y_val = Y_train_val_scaled[train_idx], Y_train_val_scaled[val_idx]
 
 
-      except Exception as e:
-                print(f"Error in {model_name} training: {e}")
-                continue
+#       # Find the best params for the current model
+#       print(f"Performing GridSearchCV for {model_name} on fold {fold + 1}...")
 
-    results[model_name] = {                                                       # save the best model for each model type, the results and the parameters
-        'best_mse_scores': best_mse_scores,
-        'best_rmse_scores': best_rmse_scores,
-        'best_params': best_params,
-        'best_model': best_model,
-        'best_model_prediction': best_model_prediction,
-        'Y_val_best' : Y_val_best
-    }
+#       # Reduce the number of CV splits for SVR to speed up
+#       cv_splits = 2 if model_name == 'SVR' else 3
 
-  return results   
+#     #   grid_search = GridSearchCV(estimator=model, param_grid=params_grid, scoring='neg_mean_squared_error', cv=cv_splits, n_jobs=-1, verbose=1) # performs search over set of parameters
+#       from sklearn.model_selection import GridSearchCV
+#       grid_search = GridSearchCV(
+#          estimator=model,
+#          param_grid=params_grid,
+#          scoring='neg_mean_squared_error',
+#          cv=cv_splits,
+#          n_jobs=-1,
+#          verbose=1
+#         )
+
+#       try:
+#         grid_search.fit(X_train, Y_train)                                         # train the model for every combination of parameters on each training set of the fold
+#         print(f"Best parameters found for fold {fold + 1}: {grid_search.best_params_}")
+
+#         # Save the best params
+#         best_params.append(grid_search.best_params_)
+
+#         # Train the best model
+#         best_model_fold = grid_search.best_estimator_                             # the model that performs the best on the last search
+
+#         # Validate using the model
+#         print("Starting validation: ")
+#         Y_pred = best_model_fold.predict(X_val)                                   # use the best model after traind and validate
+
+#         # Calculate performance for this fold
+#         mse = mean_squared_error(Y_val, Y_pred)
+#         rmse = np.sqrt(mse)
+
+#         # Save
+#         best_mse_scores.append(mse)
+#         best_rmse_scores.append(rmse)
+
+#         print(f"End of Fold {fold + 1} - MSE: {mse:.4f}, RMSE: {rmse:.4f}")
+
+#         # Update best model if this fold's model is better
+#         if mse < best_score:
+#           best_score = mse
+#           best_model = best_model_fold
+#           best_model_prediction = Y_pred
+#           Y_val_best= Y_val
+
+
+#       except Exception as e:
+#                 print(f"Error in {model_name} training: {e}")
+#                 continue
+
+#     results[model_name] = {                                                       # save the best model for each model type, the results and the parameters
+#         'best_mse_scores': best_mse_scores,
+#         'best_rmse_scores': best_rmse_scores,
+#         'best_params': best_params,
+#         'best_model': best_model,
+#         'best_model_prediction': best_model_prediction,
+#         'Y_val_best' : Y_val_best
+#     }
+
+#   return results   
 
 # ===============================================================================
 # Ensamble
 # ===============================================================================
 # Linearly Weighted Ensemble
-def linearly_weighted_ensemble(models_results, X_test):
-    """
-    Create a linearly weighted ensemble prediction across different model types.
+# def linearly_weighted_ensemble(models_results, X_test):
+#     """
+#     Create a linearly weighted ensemble prediction across different model types.
 
-    Parameters:
-    - models_results: Dictionary containing model results
-    - X_test: Test data to make predictions on (DataFrame or NumPy array)
+#     Parameters:
+#     - models_results: Dictionary containing model results
+#     - X_test: Test data to make predictions on (DataFrame or NumPy array)
 
-    Returns:
-    - Final ensemble prediction
-    """
-    # Convert DataFrame to NumPy array if necessary
-    scaler = StandardScaler()
-    X_test = scaler.fit_transform(X_test)
-    if hasattr(X_test, 'values'):
-        X_test = X_test.values
+#     Returns:
+#     - Final ensemble prediction
+#     """
+#     # Convert DataFrame to NumPy array if necessary
+#     scaler = StandardScaler()
+#     X_test = scaler.fit_transform(X_test)
+#     if hasattr(X_test, 'values'):
+#         X_test = X_test.values
 
-    mae_values = []
-    model_predictions = []
+#     mae_values = []
+#     model_predictions = []
 
-    # Calculate Mean Absolute Error (MAE) for each model
-    for model_name, result in models_results.items():
-        mae_value = np.mean(np.abs(result['best_model_prediction'] - result['Y_val_best']))
-        mae_values.append(mae_value)
+#     # Calculate Mean Absolute Error (MAE) for each model
+#     for model_name, result in models_results.items():
+#         mae_value = np.mean(np.abs(result['best_model_prediction'] - result['Y_val_best']))
+#         mae_values.append(mae_value)
 
-        # Prepare predictions based on model type
-        if model_name.strip() == 'LSTM':
-            # Reshape for LSTM
-            X_test_lstm = prepare_lstm_data(X_test, time_steps=1)
-            model_pred = result['best_model'].predict(X_test_lstm)
-        else:
-            # For other models
-            model_pred = result['best_model'].predict(X_test)
+#         # Prepare predictions based on model type
+#         if model_name.strip() == 'LSTM':
+#             # Reshape for LSTM
+#             X_test_lstm = prepare_lstm_data(X_test, time_steps=1)
+#             model_pred = result['best_model'].predict(X_test_lstm)
+#         else:
+#             # For other models
+#             model_pred = result['best_model'].predict(X_test)
 
-        # Ensure 1D prediction
-        model_predictions.append(model_pred.reshape(-1))
+#         # Ensure 1D prediction
+#         model_predictions.append(model_pred.reshape(-1))
 
-    # Calculate inverse MAE weights
-    weights = [mae_value ** -1 for mae_value in mae_values]
-    weights = np.array(weights) / np.sum(weights)
+#     # Calculate inverse MAE weights
+#     weights = [mae_value ** -1 for mae_value in mae_values]
+#     weights = np.array(weights) / np.sum(weights)
 
-    # Compute the final ensemble prediction
-    final_prediction = np.zeros(X_test.shape[0], dtype=np.float64)
+#     # Compute the final ensemble prediction
+#     final_prediction = np.zeros(X_test.shape[0], dtype=np.float64)
 
-    # Apply weighted predictions
-    for pred, weight in zip(model_predictions, weights):
-        final_prediction += weight * pred
+#     # Apply weighted predictions
+#     for pred, weight in zip(model_predictions, weights):
+#         final_prediction += weight * pred
+    
+#     # Inverse transform to get back to original scale
+#     final_prediction = shared_target_scaler.inverse_transform(final_prediction.reshape(-1, 1)).flatten()
 
-    return final_prediction
+#     return final_prediction
 
-############### CHECK IF WE NEED THIS FOR SURE AND CANNOT USE SOMETHING ELSE #################
-### USED IN THE ENSEMBLE
-def prepare_lstm_data(X, time_steps=1, features=None):
-    """
-    Prepare data for LSTM by reshaping it to fit the expected 3D shape.
+# ############### CHECK IF WE NEED THIS FOR SURE AND CANNOT USE SOMETHING ELSE #################
+# ### USED IN THE ENSEMBLE
+# def prepare_lstm_data(X, time_steps=1, features=None):
+#     """
+#     Prepare data for LSTM by reshaping it to fit the expected 3D shape.
 
-    Parameters:
-    - X: Input data, can be 1D, 2D, or already 3D.
-    - time_steps: Number of time steps in the sequence. Default is 1 for non-time series data.
-    - features: Number of features per time step. If None, it will be inferred from the data.
+#     Parameters:
+#     - X: Input data, can be 1D, 2D, or already 3D.
+#     - time_steps: Number of time steps in the sequence. Default is 1 for non-time series data.
+#     - features: Number of features per time step. If None, it will be inferred from the data.
 
-    Returns:
-    - Reshaped data with shape (samples, time_steps, features)
-    """
-    # Convert DataFrame to NumPy array if necessary
-    if hasattr(X, 'values'):
-        X = X.values
+#     Returns:
+#     - Reshaped data with shape (samples, time_steps, features)
+#     """
+#     # Convert DataFrame to NumPy array if necessary
+#     if hasattr(X, 'values'):
+#         X = X.values
 
-    if len(X.shape) == 1:
-        if features is None:
-            features = 1  # Assuming single feature if not specified
-        X = X.reshape(-1, time_steps, features)
-    elif len(X.shape) == 2:
-        samples, cols = X.shape
-        if features is None:
-            features = cols  # All columns are considered as features if not specified
-        if cols % time_steps != 0:
-            raise ValueError(f"Number of columns ({cols}) must be evenly divisible by time_steps ({time_steps})")
-        X = X.reshape(samples, cols // features, features)
-    else:
-        raise ValueError(f"Input data must be 1D or 2D. Got {len(X.shape)}D data.")
+#     if len(X.shape) == 1:
+#         if features is None:
+#             features = 1  # Assuming single feature if not specified
+#         X = X.reshape(-1, time_steps, features)
+#     elif len(X.shape) == 2:
+#         samples, cols = X.shape
+#         if features is None:
+#             features = cols  # All columns are considered as features if not specified
+#         if cols % time_steps != 0:
+#             raise ValueError(f"Number of columns ({cols}) must be evenly divisible by time_steps ({time_steps})")
+#         X = X.reshape(samples, cols // features, features)
+#     else:
+#         raise ValueError(f"Input data must be 1D or 2D. Got {len(X.shape)}D data.")
 
-    return X
+#     return X
 
-# Equal Weights Ensemble
-def equal_weighted_ensemble(models_results, X_test):
-    """
-    Calculate an equal weighted ensemble prediction.
+# # Equal Weights Ensemble
+# def equal_weighted_ensemble(models_results, X_test):
+#     """
+#     Calculate an equal weighted ensemble prediction.
 
-    Parameters:
-    - models_results: Dictionary containing model results
-    - X_test: Test data to make predictions on (DataFrame or NumPy array)
+#     Parameters:
+#     - models_results: Dictionary containing model results
+#     - X_test: Test data to make predictions on (DataFrame or NumPy array)
 
-    Returns:
-    - Final ensemble prediction
-    """
-    # Convert DataFrame to NumPy array if necessary
-    if isinstance(X_test, pd.DataFrame) or isinstance(X_test, pd.Series):
-        X_test = X_test.values
+#     Returns:
+#     - Final ensemble prediction
+#     """
+#     # Convert DataFrame to NumPy array if necessary
+#     if isinstance(X_test, pd.DataFrame) or isinstance(X_test, pd.Series):
+#         X_test = X_test.values
 
-    scaler = StandardScaler()
-    X_test = scaler.fit_transform(X_test)
+#     scaler = StandardScaler()
+#     X_test = scaler.fit_transform(X_test)
 
-    model_predictions = []
+#     model_predictions = []
 
-    # Prepare predictions based on model type
-    for model_name, result in models_results.items():
-        if 'best_model' in result:
-            if model_name.strip() == 'LSTM':
-                # Reshape for LSTM
-                X_test_lstm = prepare_lstm_data(X_test, time_steps=1)
-                model_pred = result['best_model'].predict(X_test_lstm)
-            else:
-                # For other models
-                model_pred = result['best_model'].predict(X_test)
+#     # Prepare predictions based on model type
+#     for model_name, result in models_results.items():
+#         if 'best_model' in result:
+#             if model_name.strip() == 'LSTM':
+#                 # Reshape for LSTM
+#                 X_test_lstm = prepare_lstm_data(X_test, time_steps=1)
+#                 model_pred = result['best_model'].predict(X_test_lstm)
+#             else:
+#                 # For other models
+#                 model_pred = result['best_model'].predict(X_test)
 
-            # Ensure 1D prediction
-            model_predictions.append(model_pred.reshape(-1))
+#             # Ensure 1D prediction
+#             model_predictions.append(model_pred.reshape(-1))
 
-    if not model_predictions:
-        raise ValueError("No predictions available for ensemble methods")
+#     if not model_predictions:
+#         raise ValueError("No predictions available for ensemble methods")
 
-    # Calculate weight (equal for all models)
-    weight = 1.0 / len(model_predictions)
+#     # Calculate weight (equal for all models)
+#     weight = 1.0 / len(model_predictions)
 
-    # Compute the final ensemble prediction
-    final_prediction = np.zeros(X_test.shape[0], dtype=np.float64)
+#     # Compute the final ensemble prediction
+#     final_prediction = np.zeros(X_test.shape[0], dtype=np.float64)
 
-    # Apply weighted predictions
-    for pred in model_predictions:
-        final_prediction += weight * pred
+#     # Apply weighted predictions
+#     for pred in model_predictions:
+#         final_prediction += weight * pred
+    
+#     # Inverse transform to get back to original scale
+#     final_prediction = shared_target_scaler.inverse_transform(final_prediction.reshape(-1, 1)).flatten()
 
-    return final_prediction
+#     return final_prediction
 
-# Gradient Boosting Decision Tree Ensemble
-def gbdt_ensemble(models_results, X_train, X_test, Y_train_val):
-    """
-    Use GBDT to predict based on the predictions of base models.
+# # Gradient Boosting Decision Tree Ensemble
+# def gbdt_ensemble(models_results, X_train, X_test, Y_train_val):
+#     """
+#     Use GBDT to predict based on the predictions of base models.
 
-    :param models_results: Dictionary containing model results, with 'best_model' key for each model.
-    :param X_train: Training data features to generate meta-features for training GBDT.
-    :param X_test: Test data features for final prediction.
-    :param Y_train_val: Training labels for fitting the GBDT model.
-    :return: GBDT ensemble prediction for test data.
-    """
-    scaler = StandardScaler()
+#     :param models_results: Dictionary containing model results, with 'best_model' key for each model.
+#     :param X_train: Training data features to generate meta-features for training GBDT.
+#     :param X_test: Test data features for final prediction.
+#     :param Y_train_val: Training labels for fitting the GBDT model.
+#     :return: GBDT ensemble prediction for test data.
+#     """
+#     scaler = StandardScaler()
 
-    # Scale the data
-    X_train_scaled = scaler.fit_transform(X_train)
-    X_test_scaled = scaler.transform(X_test)  # Use transform here, not fit_transform
+#     # Scale the data
+#     X_train_scaled = scaler.fit_transform(X_train)
+#     X_test_scaled = scaler.transform(X_test)  # Use transform here, not fit_transform
 
-    # Generate meta-features for training GBDT
-    train_meta_features = []
-    test_meta_features = []
+#     # Generate meta-features for training GBDT
+#     train_meta_features = []
+#     test_meta_features = []
 
-    for model_name, result in models_results.items():
-        model = result['best_model']
-        if model_name.strip() == 'LSTM':
-            X_train_lstm = prepare_lstm_data(X_train_scaled, time_steps=1)
-            X_test_lstm = prepare_lstm_data(X_test_scaled, time_steps=1)
-            train_pred = model.predict(X_train_lstm)
-            test_pred = model.predict(X_test_lstm)
-        else:
-            train_pred = model.predict(X_train_scaled)
-            test_pred = model.predict(X_test_scaled)
+#     for model_name, result in models_results.items():
+#         model = result['best_model']
+#         if model_name.strip() == 'LSTM':
+#             X_train_lstm = prepare_lstm_data(X_train_scaled, time_steps=1)
+#             X_test_lstm = prepare_lstm_data(X_test_scaled, time_steps=1)
+#             train_pred = model.predict(X_train_lstm)
+#             test_pred = model.predict(X_test_lstm)
+#         else:
+#             train_pred = model.predict(X_train_scaled)
+#             test_pred = model.predict(X_test_scaled)
 
-        train_meta_features.append(train_pred.reshape(-1))
-        test_meta_features.append(test_pred.reshape(-1))
+#         train_meta_features.append(train_pred.reshape(-1))
+#         test_meta_features.append(test_pred.reshape(-1))
 
-    # Stack predictions as meta-features
-    X_train_meta = np.column_stack(train_meta_features)
-    X_test_meta = np.column_stack(test_meta_features)
+#     # Stack predictions as meta-features
+#     X_train_meta = np.column_stack(train_meta_features)
+#     X_test_meta = np.column_stack(test_meta_features)
 
-    # Ensure consistency in number of samples
-    if X_train_meta.shape[0] != Y_train_val.shape[0]:
-        raise ValueError(f"Shape mismatch: X_train_meta {X_train_meta.shape[0]} vs Y_train_val {Y_train_val.shape[0]}")
+#     # Ensure consistency in number of samples
+#     if X_train_meta.shape[0] != Y_train_val.shape[0]:
+#         raise ValueError(f"Shape mismatch: X_train_meta {X_train_meta.shape[0]} vs Y_train_val {Y_train_val.shape[0]}")
 
-    # Train GBDT on meta-features
-    gb_model = GradientBoostingRegressor(
-        n_estimators=200,
-        learning_rate=0.05,
-        max_depth=5,
-        min_samples_split=10,
-        min_samples_leaf=5,
-        subsample=0.8,
-        random_state=42
-    )
-    gb_model.fit(X_train_meta, Y_train_val)
+#     # Train GBDT on meta-features
+#     gb_model = GradientBoostingRegressor(
+#         n_estimators=200,
+#         learning_rate=0.05,
+#         max_depth=5,
+#         min_samples_split=10,
+#         min_samples_leaf=5,
+#         subsample=0.8,
+#         random_state=42
+#     )
+#     gb_model.fit(X_train_meta, Y_train_val)
 
-    # Predict on test meta-features
-    final_prediction = gb_model.predict(X_test_meta)
+#     # Predict on test meta-features
+#     final_prediction = gb_model.predict(X_test_meta)
 
-    return final_prediction
+#     # Inverse transform to get back to original scale
+#     final_prediction = shared_target_scaler.inverse_transform(final_prediction.reshape(-1, 1)).flatten()
 
-# ===============================================================================
-# Three Ensambles Pipeline
-# ===============================================================================
-def ensemble_pipeline(models_results, X_train, X_test, Y_train, Y_test):
-    """
-    Pipeline to apply and compare different ensemble methods.
+#     return final_prediction
 
-    :param models_results: Dictionary containing model results
-    :param X_train: Training features
-    :param X_test: Testing features
-    :param Y_train: Training labels
-    :param Y_test: Testing labels
-    :return: Dictionary with ensemble results
-    """
-    ensemble_methods = {
-        'linearly_weighted': linearly_weighted_ensemble,
-        'equal_weighted': equal_weighted_ensemble,
-        'gbdt': lambda results, x_test: gbdt_ensemble(results, X_train, x_test, Y_train)
-    }
+# # ===============================================================================
+# # Three Ensambles Pipeline
+# # ===============================================================================
+# def ensemble_pipeline(models_results, X_train, X_test, Y_train, Y_test):
+#     """
+#     Pipeline to apply and compare different ensemble methods.
 
-    results = {}
+#     :param models_results: Dictionary containing model results
+#     :param X_train: Training features
+#     :param X_test: Testing features
+#     :param Y_train: Training labels
+#     :param Y_test: Testing labels
+#     :return: Dictionary with ensemble results
+#     """
+#     ensemble_methods = {
+#         'linearly_weighted': linearly_weighted_ensemble,
+#         'equal_weighted': equal_weighted_ensemble,
+#         'gbdt': lambda results, x_test: gbdt_ensemble(results, X_train, x_test, Y_train)
+#     }
 
-    for method_name, method in ensemble_methods.items():
-        # Note: 'gbdt' method requires X_train and Y_train, others do not
-        if method_name == 'gbdt':
-            final_prediction = method(models_results, X_test)
-        else:
-            final_prediction = method(models_results, X_test)
+#     results = {}
 
-        # Calculate performance metrics
-        mse = mean_squared_error(Y_test, final_prediction)
-        rmse = np.sqrt(mse)
-        r2 = r2_score(Y_test, final_prediction)
+#     for method_name, method in ensemble_methods.items():
+#         # Note: 'gbdt' method requires X_train and Y_train, others do not
+#         if method_name == 'gbdt':
+#             final_prediction = method(models_results, X_test)
+#         else:
+#             final_prediction = method(models_results, X_test)
 
-        results[method_name] = {
-            'prediction': final_prediction,
-            'mse': mse,
-            'rmse': rmse,
-            'r2': r2
-        }
+#         # Calculate performance metrics against unscaled targets
+#         # The predictions are already inverse transformed in the ensemble methods
+#         mse = mean_squared_error(Y_test, final_prediction)
+#         rmse = np.sqrt(mse)
+#         r2 = r2_score(Y_test, final_prediction)
 
-    for method_name, metrics in results.items():
-        print(f"\n{method_name} Results:")
-        print(f"  RMSE: {metrics['rmse']:.6f}")
-        print(f"  MSE: {metrics['mse']:.6f}")
-        print(f"  R2: {metrics['r2']:.6f}")
+#         results[method_name] = {
+#             'prediction': final_prediction,
+#             'mse': mse,
+#             'rmse': rmse,
+#             'r2': r2
+#         }
 
-    # Find the best performing method based on RMSE
-    best_method = min(results.items(), key=lambda x: x[1]['rmse'])[0]
-    print(f"\nBest Ensemble Method: {best_method}")
-    print(f"Performance of {best_method}:")
-    print(f"  MSE: {results[best_method]['mse']}")
-    print(f"  RMSE: {results[best_method]['rmse']}")
-    print(f"  R2: {results[best_method]['r2']}")
+#     for method_name, metrics in results.items():
+#         print(f"\n{method_name} Results:")
+#         print(f"  RMSE: {metrics['rmse']:.6f}")
+#         print(f"  MSE: {metrics['mse']:.6f}")
+#         print(f"  R2: {metrics['r2']:.6f}")
 
-    return results
+#     # Find the best performing method based on RMSE
+#     best_method = min(results.items(), key=lambda x: x[1]['rmse'])[0]
+#     print(f"\nBest Ensemble Method: {best_method}")
+#     print(f"Performance of {best_method}:")
+#     print(f"  MSE: {results[best_method]['mse']}")
+#     print(f"  RMSE: {results[best_method]['rmse']}")
+#     print(f"  R2: {results[best_method]['r2']}")
+
+#     return results
 
 # ===============================================================================
 # Full Pipeline For Single Stock
 # ===============================================================================
 def full_pipeline_for_single_stock(ticker_symbol, start_date, end_date, risk_free_rate = 0.02):
   # run first pipeline, fetch data
+  from Data_Cleaning_Pipelines import create_stock_data_pipeline, create_data_cleaning_pipeline
   pipeline = create_stock_data_pipeline(ticker_symbol, start_date, end_date, risk_free_rate)
   data = pipeline.fit_transform(pd.DataFrame())
   data.to_csv(f'{date_folder}/{ticker_symbol}_data.csv')
@@ -937,15 +638,42 @@ def full_pipeline_for_single_stock(ticker_symbol, start_date, end_date, risk_fre
 
   X_train_val = X.iloc[:split_idx]
   Y_train_val = Y.iloc[:split_idx]
-
   X_test = X.iloc[split_idx:]
   Y_test = Y.iloc[split_idx:]
 
-  results = train_and_validate_models(X_train_val, Y_train_val)
+  """"""
+
+  # TODO -> NEW PART: OPTIMIZATION: FEATURE SELECTION!!
+  from Feature_Selection_and_Optimization import analyze_feature_importance, evaluate_feature_sets
+  # Analyze feature importance
+  importance_df = analyze_feature_importance(X_train_val, Y_train_val)
+  print("Feature Importance:")
+  print(importance_df.head(10))
+  
+  # Evaluate different feature sets
+  results = evaluate_feature_sets(X_train_val, Y_train_val, X_test, Y_test)
+  print("\nFeature Selection Method Comparison:")
+  print(results[['Method', 'Num_Features', 'MSE', 'RMSE', 'R2']])
+
+  # Select best feature set based on results
+  best_method = results.loc[results['RMSE'].idxmin(), 'Method']
+  best_features = results.loc[results['RMSE'].idxmin(), 'Features']
+  print(f"\nBest feature set: {best_method} with {len(best_features)} features")
+
+  # Use these features for model training
+  X_train_val_selected = X_train_val[best_features]
+  X_test_selected = X_test[best_features]
+
+  from Models_Creation_and_Training import train_and_validate_models
+
+  # Train models using the selected features
+  results = train_and_validate_models(X_train_val_selected, Y_train_val)
+
+  from Ensembles import ensemble_pipeline
 
   # Ensamble -> Three ensambles pipeline (Linearly Weighted, Equal Weights, GBDT)
   # Run the pipeline
-  ensemble_results = ensemble_pipeline(results, X_train_val, X_test, Y_train_val, Y_test)
+  ensemble_results = ensemble_pipeline(results, X_train_val_selected, X_test_selected, Y_train_val, Y_test)
 
   # Convert dictionary to DataFrame and save to CSV
   df = pd.DataFrame.from_dict(ensemble_results, orient='index')
@@ -970,36 +698,7 @@ def full_pipeline_for_single_stock(ticker_symbol, start_date, end_date, risk_fre
   ##results_df.to_csv(f'{ticker_symbol}_ensamble_prediction_results.csv')
   results_df.to_csv(f'{date_folder}/{ticker_symbol}_ensamble_prediction_results.csv')
 
-# ===============================================================================
-# Final loop - call the pipeline for each ticker symbol
-# ===============================================================================
-# stakeholder_data = pd.read_csv('final_tickers_score.csv')
-# top_10 = stakeholder_data.head(15)['Ticker'].tolist()
-# last_10 = stakeholder_data.tail(10)['Ticker'].tolist()
 
-# def is_valid_ticker(ticker):
-#   try:
-#     ticker_data = yf.Ticker(ticker)
-#     if not ticker_data.info or 'symbol' not in ticker_data.info:
-#       print(f"Ticker {ticker} is not valid.")
-#       return False
-#     return True
-#   except Exception as e:
-#     print(f"Error validating ticker {ticker}: {e}")
-#     return False
-
-# def get_top_valid_tickers(tickers, top_n=20):
-#   valid_tickers = []
-#   for ticker in tickers:
-#     if is_valid_ticker(ticker):
-#       valid_tickers.append(ticker)
-#     if len(valid_tickers) == top_n:
-#       break
-
-#   return valid_tickers
-
-# top_10_valid_tickers = get_top_valid_tickers(top_10, top_n=10)
-
-full_pipeline_for_single_stock('OPK', "2020-01-01", "2024-01-01")
+full_pipeline_for_single_stock('OPK', "2023-12-01", "2024-01-01")
